Add files via upload
Here I have used  Keras to build and train a neural network that analyzes text for sentiment. here i have used IMDB movie reviews sentiment
classifiaction dataet that's included with keras. The sentiment that's expressed in these reviews is the basis for which your neural network will analyze text presented to it and score it for sentiment.